Export toggleSidebar
