+++ src/app/header/header.component.html
@@ -13,8 +13,8 @@
         </a>
       </li>
       <li routerLinkActive="active" [routerLinkActiveOptions]="{exact: true}">
-        <a routerLink="/about">
-          <i class="glyphicon glyphicon-info-sign"></i> About
+        <a class="pointer" (click)="toggleSidebar()">
+          <i class="glyphicon glyphicon-list"></i> Menu
         </a>
       </li>
     </ul>
