+++ src/app/header/sidebar/sidebar.component.html
@@ -53,7 +53,7 @@
   </div>
 </div>
 <div class="sidebar-toggle" id="sidebar-toggle">
-  <a (click)="toggleSidebar()">
+  <a (click)="head.toggleSidebar()">
     <i class="pe-7s-menu sidebar-toggle-icon" id="sidebar-toggle-icon"></i>
   </a>
 </div>
